requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
motor==3.3.2
pymongo==4.6.1

# Scraping
playwright==1.41.0
beautifulsoup4==4.12.3
lxml==5.1.0
aiohttp==3.9.1

# PDF Processing
PyMuPDF>=1.24.0

# LLM - Gemini
google-generativeai==0.3.2

# Embeddings
sentence-transformers==2.2.2
numpy==1.26.3

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
tenacity==8.2.3
aiofiles==23.2.1

# Scheduling
apscheduler==3.10.4

# Logging
structlog==24.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
